Accessory, internal functions for taxonomic manipulation

These accessory functions work for editing the notation of scientific names and are mainly used within plantR function fixSpecies().

rmOpen(x)

rmInfra(x)

rmHyb(x)

addRank(x, rank = NULL)

Arguments

x
a vector with scientific names to be standardized.

rank
the expression or symbol to be added between names.

Details

The functions rmOpen(), rmInfra(), and rmHyb() require only a vector of scientific names, while the function addRank() also requires a rank to be provided.

The function rmOpen() removes the open nomenclature 'cf.' and 'aff.'.

The function rmInfra() removes the infra-specific ranks from varieties, sub-species and forms (e.g. 'var.', 'subsp.', 'f.')

The function rmHyb() removes the hybrid symbol 'x'.

The function addRank() does the opposite operation: it adds ranks, open nomenclature or hybrid symbols into scientific names. For this function if the number of ranks is equal to the number of names provided each rank is assigned to the corresponding name. If the number of ranks is different, the function silently uses the first rank for all names.

See also

Author

Sara Mortara & Renato A. F. de Lima

Examples

if (FALSE) {
rmOpen(c("Lindsaea cf. lancea","Lindsaea aff. lancea"))
}

if (FALSE) {
rmInfra(c("Lindsaea lancea var. angulata",
          "Lindsaea lancea (L.) Bedd. var. angulata Rosenst.",
          "Lindsaea schomburgkii f. coriifolia (Lindm.) K.U. Kramer"))
}

if (FALSE) {
rmHyb(c("Blechnum ×antillanum", "Blechnum × antillanum",
       "Blechnum x antillanum", "Blechnum X antillanum"))
}

if (FALSE) {
addRank(c("Lindsaea lancea", "Lindsaea lancea angulata",
          "Lindsaea schomburgkii coriifolia", "Blechnum antillanum"),
          rank = c("cf.", "var.", "f.", "\u00d7"))
}