Create Species (Check-)List

This function creates a list of the species contained in the occurrence data, including a list of voucher specimens. If the occurrence data is associated to an specific geographical area, this list can be used as an inventory or check-list for the species occurring in that area.

checkList(
  x,
  fam.order = TRUE,
  n.vouch = 30,
  type = "short",
  rm.dup = TRUE,
  rank.type = 5,
  date.format = "%d %b %Y"
)

Arguments

x
a data frame with the occurrence data, generally as the output of the plantR validation functions.

fam.order
logical. Should taxa be organized in alphabetical order of families? Defaults to TRUE.

n.vouch
numerical. Maximum number of vouchers to be listed per taxa. Defaults to 30.

type
character. The type of voucher list desired. Options are: 'short', 'selected' and 'list' (see details below).

rm.dup
logical. Should duplicated specimens be removed prior to the calculation of species summaries? Defaults to TRUE.

rank.type
numerical. Value of the ranking for type specimens in order to organize order and filter the voucher list. Defaults to 5.

date.format
The desired format for the dates. Defaults to "%d %b %Y"

Details

The list of species can be organized in alphabetic order by taxa or in alphabetic order by family and then by taxa within families (the default).

By default, the output list provides the number of records found and the overall taxonomic and geographic confidence level of the records (columns 'tax.CL' and 'geo.CL'), if available. The taxonomic confidence level is the percentage of records with the identification flagged as 'high', while the geographic confidence level is the percentage of records with coordinates flagged as being validated at municipality or locality levels.

The function also provides a list of vouchers, giving priority to type specimens and records with higher level of confidence in their identification. By default, the function provides up to 30 vouchers per taxa, but this number can be controlled using the argument n.vouch.

The voucher list can be provided in the following output formats (the option 'list' is not implemented yet):

'short': Collector name, Collector number (collections of deposit)

'selected': COUNTRY, stateProvince: municipality, Date, Collector name, Collector number (collections of deposit)

'list': Collector name, Collector number(s) (species code)

Note 1: although we provide a date.format argument, checks of other date formats other than the default are pending and so they may not work properly.

Note 2: The columns names of the input data are expected to be in the DarwinCore format or in the standard output names of the plantR workflow. Currently, there is no argument to make the equivalency to different column names, so users need to convert their data into one of these two options. See function formatDwc() for more details.

Examples

(df <- data.frame(collectionCode = c("CRI","CRI","CRI","CRI"),
catalogNumber = c("3565","713","3073","15331"),
recordedBy = c("Rebelo, M.C.","Citadini-Zanette, V.",
"Santos, R.","Zapelini, I."),
recordNumber = c("s.n.","1063","11","s.n."),
year = c("1994","1990","1994","2020"),
family = c("Salicaceae","Salicaceae","Cannabaceae","Cannabaceae"),
scientificName = c("Casearia sylvestris","Casearia sylvestris",
"Trema micrantha","Trema micrantha"),
country = c("brazil","brazil","brazil","brazil"),
stateProvince = c("santa catarina","santa catarina",
"santa catarina","santa catarina"),
municipality = c("jaguaruna","orleans","icara",NA)))

#>   collectionCode catalogNumber           recordedBy recordNumber year
#> 1            CRI          3565         Rebelo, M.C.         s.n. 1994
#> 2            CRI           713 Citadini-Zanette, V.         1063 1990
#> 3            CRI          3073           Santos, R.           11 1994
#> 4            CRI         15331         Zapelini, I.         s.n. 2020
#>        family      scientificName country  stateProvince municipality
#> 1  Salicaceae Casearia sylvestris  brazil santa catarina    jaguaruna
#> 2  Salicaceae Casearia sylvestris  brazil santa catarina      orleans
#> 3 Cannabaceae     Trema micrantha  brazil santa catarina        icara
#> 4 Cannabaceae     Trema micrantha  brazil santa catarina         <NA>

checkList(df, rm.dup = FALSE)

#>        family      scientificName records
#> 2 Cannabaceae     Trema micrantha       2
#> 1  Salicaceae Casearia sylvestris       2
#>                                                                vouchers
#> 2           Santos, R., 11 (CRI, 3073); Zapelini, I., s.n. (CRI, 15331)
#> 1 Citadini-Zanette, V., 1063 (CRI, 713); Rebelo, M.C., s.n. (CRI, 3565)

checkList(df, rm.dup = FALSE, type = "selected")

#>        family      scientificName records
#> 2 Cannabaceae     Trema micrantha       2
#> 1  Salicaceae Casearia sylvestris       2
#>                                                                                                                                     vouchers
#> 2                          BRAZIL, Santa Catarina, 2020, Zapelini, I., s.n. (CRI); BRAZIL, Santa Catarina: Içara, 1994, Santos, R., 11 (CRI)
#> 1 BRAZIL, Santa Catarina: Jaguaruna, 1994, Rebelo, M.C., s.n. (CRI); BRAZIL, Santa Catarina: Orleaes, 1990, Citadini-Zanette, V., 1063 (CRI)