Format Locality Information

Standardize the notation of the locality fields country, stateProvince, municipality and locality, and search from some missing information within the available locality information.

fixLoc(
  x,
  loc.levels = c("country", "stateProvince", "municipality", "locality"),
  scrap = TRUE,
  to.lower = TRUE
)

Arguments

x
a data frame containing typical locality fields from species records.

loc.levels
a vector containing the names of the locality fields to be formatted.

scrap
logical. Should the search of missing locality information be performed? Default to TRUE.

to.lower
logical. Should the output locality names be return in lower cases? Default to TRUE.

Value

The input data frame x, plus the '.new' columns with the formatted fields and the resolution of the locality information available.

Details

The function performs several edits and replacements. Country names are formatted into the international format, letters are lower-cased, and special characters and common abbreviations are removed.

By default, this function formats all four locality fields simultaneously (i.e. country, stateProvince, municipality, locality), but the user can choose among these fields through the argument loc.levels. However, the process of searching for missing information is more complete if all the four locality fields mentioned above are available.

If present, other Darwin Core fields are used internally to obtain missing information on the locality fields declared above, namely: 'countryCode', 'county', and 'verbatimLocality'.

The argument scrap controls the search for missing municipality information from the field 'locality'. It also performs some extra editing and cropping of the field 'locality' in order to obtain more standardized locality descriptions. This argument uses different ways of splitting and cropping the locality description in order to find missing information. Although it does not always result in an accurate extraction of the information, it provides an extra tool to organize locality information which are not provided in the appropriate columns.

The function automatically returns the original resolution of the locality information provided. For instance, if only country information is provided (i.e. field is not empty), then the resolution is flagged as 'country'; if country and stateProvince are given, then the resolution is flagged as 'stateProvince', and so on.

Author

Renato A. F. de Lima

Examples

# Creating a data frame with locality information
(df <- data.frame(country = c("BR", "Brasil", "BRA", "Brazil", NA),
stateProvince = c("MG", "estado de Minas Gerais", "Minas Geraes",
"Minas Gerais", "Minas Gerais"),
municipality = c("Lavras", "lavras", NA, NA, "Lavras"),
locality = c(NA, "UFLA", "municipio de Lavras, campus UFLA",
"Minas Gerais, municipio Lavras", NA)))

#>   country          stateProvince municipality                         locality
#> 1      BR                     MG       Lavras                             <NA>
#> 2  Brasil estado de Minas Gerais       lavras                             UFLA
#> 3     BRA           Minas Geraes         <NA> municipio de Lavras, campus UFLA
#> 4  Brazil           Minas Gerais         <NA>   Minas Gerais, municipio Lavras
#> 5    <NA>           Minas Gerais       Lavras                             <NA>

# Formating the locality information
fixLoc(df, scrap = FALSE)

#>   country          stateProvince municipality                         locality
#> 1      BR                     MG       Lavras                             <NA>
#> 2  Brasil estado de Minas Gerais       lavras                             UFLA
#> 3     BRA           Minas Geraes         <NA> municipio de Lavras, campus UFLA
#> 4  Brazil           Minas Gerais         <NA>   Minas Gerais, municipio Lavras
#> 5    <NA>           Minas Gerais       Lavras                             <NA>
#>   country.new stateProvince.new municipality.new
#> 1      brazil      minas gerais           lavras
#> 2      brazil      minas gerais           lavras
#> 3      brazil      minas gerais             <NA>
#> 4      brazil      minas gerais             <NA>
#> 5      brazil      minas gerais           lavras
#>                       locality.new    resol.orig
#> 1                             <NA>  municipality
#> 2                             ufla      locality
#> 3 municipio de lavras, campus ufla stateProvince
#> 4   minas gerais, municipio lavras stateProvince
#> 5                             <NA>  municipality

fixLoc(df, scrap = FALSE, to.lower = FALSE)

#>   country          stateProvince municipality                         locality
#> 1      BR                     MG       Lavras                             <NA>
#> 2  Brasil estado de Minas Gerais       lavras                             UFLA
#> 3     BRA           Minas Geraes         <NA> municipio de Lavras, campus UFLA
#> 4  Brazil           Minas Gerais         <NA>   Minas Gerais, municipio Lavras
#> 5    <NA>           Minas Gerais       Lavras                             <NA>
#>   country.new stateProvince.new municipality.new
#> 1      Brazil      Minas Gerais           Lavras
#> 2      Brazil      Minas Gerais           Lavras
#> 3      Brazil      Minas Gerais             <NA>
#> 4      Brazil      Minas Gerais             <NA>
#> 5      Brazil      Minas Gerais           Lavras
#>                       locality.new    resol.orig
#> 1                             <NA>  municipality
#> 2                             Ufla      locality
#> 3 Municipio De Lavras, Campus Ufla stateProvince
#> 4   Minas Gerais, Municipio Lavras stateProvince
#> 5                             <NA>  municipality

# Formating and scrapping the locality information
fixLoc(df, scrap = TRUE)

#>   country          stateProvince municipality                         locality
#> 1      BR                     MG       Lavras                             <NA>
#> 2  Brasil estado de Minas Gerais       lavras                             UFLA
#> 3     BRA           Minas Geraes         <NA> municipio de Lavras, campus UFLA
#> 4  Brazil           Minas Gerais         <NA>   Minas Gerais, municipio Lavras
#> 5    <NA>           Minas Gerais       Lavras                             <NA>
#>   country.new stateProvince.new municipality.new locality.new
#> 1      brazil      minas gerais           lavras         <NA>
#> 2      brazil      minas gerais           lavras         ufla
#> 3      brazil      minas gerais           lavras  campus ufla
#> 4      brazil      minas gerais           lavras minas gerais
#> 5      brazil      minas gerais           lavras         <NA>
#>        locality.scrap   resol.orig
#> 1                <NA> municipality
#> 2                ufla     locality
#> 3 municipio de lavras     locality
#> 4        minas gerais     locality
#> 5                <NA> municipality

# Formating the locality information only at country and state levels
fixLoc(df, loc.levels = c("country", "stateProvince"))[,-c(1:4)]

#>   country.new stateProvince.new    resol.orig
#> 1      brazil      minas gerais stateProvince
#> 2      brazil      minas gerais stateProvince
#> 3      brazil      minas gerais stateProvince
#> 4      brazil      minas gerais stateProvince
#> 5        <NA>      minas gerais       no_info