Edit Scientific Name Notation

Identifies open nomenclature (aff., cf.) in scientific names, classification under species level (var. and subsp.). It creates a new column with the new suggested name and it also flags problematic names (character string with numbers, authors, wrong case, or other names besides genus and epithet etc). Names can be returned with or without infra-specific ranks (var. and subsp.) or abbreviations of unspecific names (sp. or spp.). In the case of names with authors, authorship is currently removed from scientific names.

fixSpecies(
  x = NULL,
  tax.name = "scientificName",
  rm.rank = FALSE,
  rm.indet = FALSE
)

Arguments

x
a vector or data.frame containing the species name

tax.name
character. Name of the columns containing the species name. Default to "scientificName"

rm.rank
logical. Should the infra-specific rank abbreviation be removed from the name? Default to FALSE

rm.indet
logical. Should the abbreviations for unspecific names (i.e. sp. or spp.) be removed? Default to FALSE

Value

The original data frame (or the input vector as a data frame) with the new columns verbatimSpecies with small edits before flagging, scientificNameStatus with the flags in original data and scientificName.new with a suggestion for a more correct name. See Details for a description of flags in the column scientificNameStatus.

Details

Possible flags returned in scientificNameStatus:

possibly_ok

scientific name following the expected pattern 'Genus epithet'

not_Genus_epithet_format

scientific name not following the expected pattern Genus epithet

variety

scientific name with variety

subspecies

scientific name with subspecies

form

scientific name with form

infra_specific

scientific name with genus, specific epiteth and infra-specific, but no infra-specific rank

hybrid_species

scientific name of a hybrid species

conferre

open nomenclature cf. in the scientific name

affinis

open nomenclature aff. in the scientific name

indet

taxon identified only at genus level

subfamily_as_genus

subfamily as genus, not a valid name

family_as_genus

family as genus, not a valid name

order_as_genus

order as genus, not a valid name

incertae_sedis

scientific name of uncertain placement

species_nova

species name contains an indication of a new species, possibly not yet a valid name

name_w_authors

scientific name has authors

name_w_wrong_case

scientific name has upper/lowercase issues

name_w_non_ascii

species name has non ASCII characters, not a valid name

abbreviated_genus

genus is abbreviated

not_name_has_digits

scientific name has digits, not a valid name

References

Sigovini, M., Keppel, E. and Tagliapietra, D. (2016) Open Nomenclature in the biodiversity era. Methods in Ecology and Evolution 7(10): 1217-1225.

Author

Sara Mortara & Renato A. Ferreira de Lima

Examples

df <- data.frame(scientificName =
c("Lindsaea lancea", "Lindsaea lancea (L.) Bedd.",
"Lindsaea lancea var. Angulata",
"Lindsaea Aff. lancea",
"Lindsaea", "Lindsaea sp.", "Lindsaeaceae sp.",
"Lindsaea aff. lancea (L.) Bedd.",
"Lindsaea ×improvisa K.U.Kramer",
"Parablechnum C.Presl",
"Blechnum spannagelii Rosenst.",
"Blechnum occidentale leopoldense Dutra",
"Blechnum austrobrasilianum de la Sota"))

fixSpecies(df)

#>                            scientificName               scientificName.new
#> 1                         Lindsaea lancea                  Lindsaea lancea
#> 2              Lindsaea lancea (L.) Bedd.                  Lindsaea lancea
#> 3           Lindsaea lancea var. Angulata    Lindsaea lancea var. angulata
#> 4                    Lindsaea Aff. lancea                  Lindsaea lancea
#> 5                                Lindsaea                     Lindsaea sp.
#> 6                            Lindsaea sp.                     Lindsaea sp.
#> 7                        Lindsaeaceae sp.                 Lindsaeaceae sp.
#> 8         Lindsaea aff. lancea (L.) Bedd.                  Lindsaea lancea
#> 9          Lindsaea ×improvisa K.U.Kramer              Lindsaea ×improvisa
#> 10                   Parablechnum C.Presl                     Parablechnum
#> 11          Blechnum spannagelii Rosenst.             Blechnum spannagelii
#> 12 Blechnum occidentale leopoldense Dutra Blechnum occidentale leopoldense
#> 13  Blechnum austrobrasilianum de la Sota       Blechnum austrobrasilianum
#>             scientificNameStatus
#> 1                    possibly_ok
#> 2                 name_w_authors
#> 3      variety|name_w_wrong_case
#> 4                        affinis
#> 5                          indet
#> 6                          indet
#> 7                family_as_genus
#> 8         affinis|name_w_authors
#> 9  hybrid_species|name_w_authors
#> 10          indet|name_w_authors
#> 11                name_w_authors
#> 12 infra_specific|name_w_authors
#> 13                name_w_authors

fixSpecies(df, rm.rank = TRUE)

#>                            scientificName               scientificName.new
#> 1                         Lindsaea lancea                  Lindsaea lancea
#> 2              Lindsaea lancea (L.) Bedd.                  Lindsaea lancea
#> 3           Lindsaea lancea var. Angulata         Lindsaea lancea angulata
#> 4                    Lindsaea Aff. lancea                  Lindsaea lancea
#> 5                                Lindsaea                     Lindsaea sp.
#> 6                            Lindsaea sp.                     Lindsaea sp.
#> 7                        Lindsaeaceae sp.                 Lindsaeaceae sp.
#> 8         Lindsaea aff. lancea (L.) Bedd.                  Lindsaea lancea
#> 9          Lindsaea ×improvisa K.U.Kramer               Lindsaea improvisa
#> 10                   Parablechnum C.Presl                     Parablechnum
#> 11          Blechnum spannagelii Rosenst.             Blechnum spannagelii
#> 12 Blechnum occidentale leopoldense Dutra Blechnum occidentale leopoldense
#> 13  Blechnum austrobrasilianum de la Sota       Blechnum austrobrasilianum
#>             scientificNameStatus
#> 1                    possibly_ok
#> 2                 name_w_authors
#> 3      variety|name_w_wrong_case
#> 4                        affinis
#> 5                          indet
#> 6                          indet
#> 7                family_as_genus
#> 8         affinis|name_w_authors
#> 9  hybrid_species|name_w_authors
#> 10          indet|name_w_authors
#> 11                name_w_authors
#> 12 infra_specific|name_w_authors
#> 13                name_w_authors

fixSpecies(df, rm.rank = TRUE, rm.indet = TRUE)

#>                            scientificName               scientificName.new
#> 1                         Lindsaea lancea                  Lindsaea lancea
#> 2              Lindsaea lancea (L.) Bedd.                  Lindsaea lancea
#> 3           Lindsaea lancea var. Angulata         Lindsaea lancea angulata
#> 4                    Lindsaea Aff. lancea                  Lindsaea lancea
#> 5                                Lindsaea                         Lindsaea
#> 6                            Lindsaea sp.                         Lindsaea
#> 7                        Lindsaeaceae sp.                     Lindsaeaceae
#> 8         Lindsaea aff. lancea (L.) Bedd.                  Lindsaea lancea
#> 9          Lindsaea ×improvisa K.U.Kramer               Lindsaea improvisa
#> 10                   Parablechnum C.Presl                     Parablechnum
#> 11          Blechnum spannagelii Rosenst.             Blechnum spannagelii
#> 12 Blechnum occidentale leopoldense Dutra Blechnum occidentale leopoldense
#> 13  Blechnum austrobrasilianum de la Sota       Blechnum austrobrasilianum
#>             scientificNameStatus
#> 1                    possibly_ok
#> 2                 name_w_authors
#> 3      variety|name_w_wrong_case
#> 4                        affinis
#> 5                          indet
#> 6                          indet
#> 7                family_as_genus
#> 8         affinis|name_w_authors
#> 9  hybrid_species|name_w_authors
#> 10          indet|name_w_authors
#> 11                name_w_authors
#> 12 infra_specific|name_w_authors
#> 13                name_w_authors