Flag Occurrences From Cultivated Individuals

This function detects records from cultivated individuals, based on record locality, description and habitat.

getCult(
  x,
  remarks = "occurrenceRemarks",
  loc.name = "locality",
  habitat = "habitat"
)

Arguments

x
a data frame with the species record data.

remarks
character. The name of the column containing the record remarks. Default to the Darwin Core standard 'occurrenceRemarks'.

loc.name
character. The name of the column containing the record locality information. Default to 'locality'.

habitat
character. The name of the column containing the record habitat information. Default to 'habitat'.

Value

The input data frame with an additional column 'cult.check' with the result of the search for records from cultivated individuals.

Details

The input data frame x should contain at least one of the columns containing the description of the record locality (e.g. 'locality'), remarks (e.g. 'occurrenceRemarks'), or habitat. The names of columns in which these information is stored can de declared using the arguments loc.name, remarks and habitat (defaults to the Darwin Core standard notation).

If present, other Darwin Core fields are used internally to obtain missing information on the three fields declared above, namely: 'verbatimLocality', 'biologicalStatus' and 'fieldNotes'. Therefore, the search for cultivated individuals in plantR is NOT based on the geographical coordinates of the species records, although records flagged as spatial outliers by plantR (see function checkOut()) may also be flagged as cultivated.

The search of records from cultivated individuals is performed on all the fields available and it is based on a list of terms that denotes clear indication of cultivated individuals (e.g. 'Cultivated', 'Planted', 'Exotic'). The function returns the column 'cult.check' with two categories:

"cultivated": exact matches of the list of terms of cultivated individuals with the text in at least one of the fields mentioned above.

"prob_cultivated": presence of one or more terms in the fields mentioned above.

For assigning the "prob_cultivated", a second list of terms is used to exclude possible spurious hits of cultivated individuals (e.g 'Cultivated area' or 'Presence of exotic species'). But this list is not extensive and so this category may need some level of double-checking by the user.

Examples

(df = data.frame(
  occurrenceRemarks = c("tree, 10 m", "Frutos Roxos. Cultivada", NA, "Tree"),
  locality = c("pastagem cultivada", NA, "Itatiaia, cultivada perto da sede", "Brazil"),
  habitat = c(NA, "Floresta", "Mata", "Cultivated"), stringsAsFactors = FALSE))

#>         occurrenceRemarks                          locality    habitat
#> 1              tree, 10 m                pastagem cultivada       <NA>
#> 2 Frutos Roxos. Cultivada                              <NA>   Floresta
#> 3                    <NA> Itatiaia, cultivada perto da sede       Mata
#> 4                    Tree                            Brazil Cultivated

getCult(df)

#>         occurrenceRemarks                          locality    habitat
#> 1              tree, 10 m                pastagem cultivada       <NA>
#> 2 Frutos Roxos. Cultivada                              <NA>   Floresta
#> 3                    <NA> Itatiaia, cultivada perto da sede       Mata
#> 4                    Tree                            Brazil Cultivated
#>        cult.check
#> 1            <NA>
#> 2 prob_cultivated
#> 3 prob_cultivated
#> 4      cultivated