Standardize Botanical Family

Search for valid family names of vascular plants following the APG IV (2016) and in the PPG I (2016).

prepFamily(
  x,
  fam.name = "family",
  gen.name = "genus",
  spp.name = "scientificName",
  print = TRUE
)

Arguments

x
a data frame containing the names of the vascular plant families and the corresponding species.

fam.name
character. The name of the column containing the family names. Default to "family".

gen.name
character. The name of the column containing the genus names. Default to "genus"

spp.name
character. The name of the column containing the species names. Default to "scientificName"

print
logical. Should the automatically replaced family names be printed? Default to TRUE.

Value

the data frame x with an additional column called 'family.new'.

Details

The data frame x must contain at least two columns which stores the family and species names. By default, the names of these columns should be family and speciesName, following Darwin Core field names. Optionally, a column with the genus name can be provided (default column name: 'genus'). If it is not provided, this columns is generated internally based on the species names provided.

The first search of names is based on the list of families and accepted synonyms for vascular plants provided by plantR, which was mainly compiled from the APG website, which includes families cited in the APG IV (2016) and in the PPG I (2016).

If the family name is not found in this list, a second search is carried in the Brazilian Flora 2020 (BF-2020), using package flora (Carvalho 2019). If the family name is still not found in BF-2020, then a final try is performed in The Plant List (TPL), using package Taxonstand (Cayuela et al. 2017). Family names retrieved from BF-2020 and TPL are finally converted to the names accepted by the APG IV or PPG I.

In case there is a conflict in the original family name and the name found based on the genus name, the original name is replaced by the name from the APG IV or PPG I with a warning.

References

Angiosperm Phylogeny Group (2016). An update of the Angiosperm Phylogeny Group classification for the orders and families of flowering plants: APG IV. Bot. J. Linnean Soc. 181: 1-20.

Gustavo Carvalho (2019). flora: Tools for Interacting with the Brazilian Flora 2020. R package version 0.3.1. http://www.github.com/gustavobio/flora

Luis Cayuela, Anke Stein and Jari Oksanen (2017). Taxonstand: Taxonomic Standardization of Plant Species Names. R package version 2.1. https://CRAN.R-project.org/package=Taxonstand

PPG I (2016). A community-derived classification for extant lycophytes and ferns. Journal of Systematics and Evolution. 54 (6): 563–603.

Author

Renato A. F. de Lima

Examples

df <- data.frame(
family = c("Ulmaceae", "Cannabaceae", "Salicaceae", "Flacourtiaceae", "Vivianiaceae", ""),
genus = c("Trema", "Trema", "Casearia", "Casearia", "Casearia", ""),
speciesName = c("Trema micrantha", "Trema micrantha", "Casearia sylvestris",
"Casearia sylvestris","Casearia sylvestris","Casearia sylvestris"),
stringsAsFactors = FALSE)
prepFamily(df, spp.name = "speciesName")

#> The following family names were automatically replaced:
#> 
#> |Genus    |Old fam.     |New fam.    |
#> |:--------|:------------|:-----------|
#> |Casearia |Vivianiaceae |Salicaceae  |
#> |Trema    |Ulmaceae     |Cannabaceae |
#> 

#>           family    genus         speciesName  family.new
#> 1       Ulmaceae    Trema     Trema micrantha Cannabaceae
#> 2    Cannabaceae    Trema     Trema micrantha Cannabaceae
#> 3     Salicaceae Casearia Casearia sylvestris  Salicaceae
#> 4 Flacourtiaceae Casearia Casearia sylvestris  Salicaceae
#> 5   Vivianiaceae Casearia Casearia sylvestris  Salicaceae
#> 6           <NA> Casearia Casearia sylvestris  Salicaceae