Format Locality String

Simplify the notation of the standard plantR locality string, which is used for the validate localities and geographical coordinates.

prepLoc(x)

Arguments

x
character. The concatenated locality string to be simplified.

Value

The edited locality string.

Details

The goal of this function is to simplify the locality strings in order to reduce variations in locality notation and thus to increase the chances of match with the plantR gazetteer and maps.

It transforms the string to lowercase and removes special characters and prepositions of locality names.

See also

Author

Renato A. F. de Lima

Examples

  prepLoc("Brazil_Rio de Janeiro")

#> [1] "brazil_rio janeiro"

  prepLoc("Brazil_Rio Grande do Sul_Coqueiros do Sul")

#> [1] "brazil_rio grande sul_coqueiros sul"

  prepLoc("Brazil_Bahia_Arraial d'Ajuda")

#> [1] "brazil_bahia_arraial ajuda"

  prepLoc("Brazil_São Paulo_Sete Barras_Parque Estadual de Carlos Botelho")

#> [1] "brazil_sao paulo_sete barras_parque estadual carlos botelho"

  prepLoc("Cuba_Isla de la Juventud")

#> [1] "cuba_isla juventud"