Get Valid Plant Names

Provides valid plant names or corrects typos and orthographic variants based on the Brazilian Flora 2020 project and on the The Plant List. It uses the functionalities of other R packages, namely flora (Carvalho 2019) and Taxonstand (Cayuela et al. 2017).

prepSpecies(
  x,
  tax.names = c("scientificName.new", "scientificNameAuthorship"),
  db = c("bfo", "tpl"),
  sug.dist = 0.9,
  use.authors = TRUE,
  drop.cols = c("tmp.ordem", "family", "verbatimSpecies", "author", "full_sp",
    "authorship", "id")
)

Arguments

x
a data.frame containing the species scientific name without authors and their authorship, ideally already passed by the string check in fixSpecies. The authorship is only used for solving homonyms.

tax.names
character. Names of the columns containing the species names and authors. Defaults to 'scientificName.new' and 'scientificNameAuthorship'.

db
the list of database to be consulted for valid names, in their preferred order of priority. Only the results from Brazilian Flora 2020 ('bfo'), The Plant List ('tpl') or both are currently implemented.

sug.dist
a fraction expressing the maximum distance allowed between the original species name and the suggested species name, which is passed to the arguments suggestion.distance of function flora::get.taxa() and max.distance of function Taxonstand::TPL(). Defaults to 0.9.

use.authors
logical. Should all authors names be verified (takes longer)? Default to TRUE.

drop.cols
character. Name of columns that should be dropped from the results.

Value

A data frame with the same columns provided by the user and some additional columns containing the suggested binomial ('suggestedName'), the corresponding scientific name ('scientific.name'), the notes and the sources used to retrieve the suggested binomial (columns 'notes' and 'source').

Details

The function currently uses packages flora and Taxonstand to provide suggested names. Not all information returned by those packages are returned here, but some of them can be controlled using the argument drop.cols.

The function does not provide the suggested names from the different databases/sources. It combines into the same column ('suggestedName') the suggestions from one or more databases. If more than one databases is used, the suggested names are combined following the same order of the database codes provided in the argument db. See examples for details.

References

Gustavo Carvalho (2019). flora: Tools for Interacting with the Brazilian Flora 2020. R package version 0.3.1. http://www.github.com/gustavobio/flora

Luis Cayuela, Anke Stein and Jari Oksanen (2017). Taxonstand: Taxonomic Standardization of Plant Species Names. R package version 2.1. https://CRAN.R-project.org/package=Taxonstand

See also

Author

Renato A. Ferreira de Lima

Examples

sp.list <- c("Casearia sylvestris","Casearia sylvestirs",
  "Casearia sylvestris angustifolia", "Casearia attenuata",
  "Casearia celtidifolia", "Casearia celtidifolia","Casearia tropicana")
aut_list <- c("Sw.", "Sw.", "Uittien", "Rusby", "Kunth", "de Vriese","L.")

df <- data.frame(scientificName.new = sp.list,
  scientificNameAuthorship = aut_list)

prepSpecies(df)

#>   |                                                          |                                                  |   0%  |                                                          |+++++++                                           |  14%  |                                                          |++++++++++++++                                    |  29%  |                                                          |+++++++++++++++++++++                             |  43%  |                                                          |+++++++++++++++++++++++++++++                     |  57%  |                                                          |++++++++++++++++++++++++++++++++++++              |  71%  |                                                          |+++++++++++++++++++++++++++++++++++++++++++       |  86%  |                                                          |++++++++++++++++++++++++++++++++++++++++++++++++++| 100%

#>                 scientificName.new scientificNameAuthorship       suggestedName
#> 1              Casearia sylvestris                      Sw. Casearia sylvestris
#> 2              Casearia sylvestirs                      Sw. Casearia sylvestris
#> 7 Casearia sylvestris angustifolia                  Uittien                <NA>
#> 3               Casearia attenuata                    Rusby Casearia sylvestris
#> 4            Casearia celtidifolia                    Kunth  Casearia ulmifolia
#> 5            Casearia celtidifolia                de Vriese  Casearia ulmifolia
#> 6               Casearia tropicana                       L.  Casearia tropicana
#>                 scientificNameFull        tax.notes tax.source
#> 1          Casearia sylvestris Sw.                    BR-Flora
#> 2          Casearia sylvestris Sw.   was misspelled   BR-Flora
#> 7                             <NA>             <NA>   BR-Flora
#> 3          Casearia sylvestris Sw. replaced synonym   BR-Flora
#> 4 Casearia ulmifolia Vahl ex Vent. replaced synonym   BR-Flora
#> 5 Casearia ulmifolia Vahl ex Vent. replaced synonym   BR-Flora
#> 6               Casearia tropicana        not found        TPL

prepSpecies(df, db = "tpl", sug.dist = 0.85)

#>   |                                                          |                                                  |   0%  |                                                          |+++++++                                           |  14%  |                                                          |++++++++++++++                                    |  29%  |                                                          |+++++++++++++++++++++                             |  43%  |                                                          |+++++++++++++++++++++++++++++                     |  57%  |                                                          |++++++++++++++++++++++++++++++++++++              |  71%  |                                                          |+++++++++++++++++++++++++++++++++++++++++++       |  86%  |                                                          |++++++++++++++++++++++++++++++++++++++++++++++++++| 100%

#>                 scientificName.new scientificNameAuthorship       suggestedName
#> 1              Casearia sylvestris                      Sw. Casearia sylvestris
#> 2              Casearia sylvestirs                      Sw. Casearia sylvestris
#> 3 Casearia sylvestris angustifolia                  Uittien Casearia sylvestris
#> 4               Casearia attenuata                    Rusby Casearia sylvestris
#> 5            Casearia celtidifolia                    Kunth  Casearia ulmifolia
#> 6            Casearia celtidifolia                de Vriese Casearia sylvestris
#> 7               Casearia tropicana                       L.  Casearia tropicana
#>                 scientificNameFull        tax.notes tax.source
#> 1          Casearia sylvestris Sw.                         TPL
#> 2          Casearia sylvestris Sw.   was misspelled        TPL
#> 3          Casearia sylvestris Sw. replaced synonym        TPL
#> 4          Casearia sylvestris Sw. replaced synonym        TPL
#> 5 Casearia ulmifolia Vahl ex Vent. replaced synonym        TPL
#> 6          Casearia sylvestris Sw. replaced synonym        TPL
#> 7               Casearia tropicana        not found        TPL