Format People's Name

Convert people's names to different formats (i.e. Last Name, First Names(s) or First Name(s), Last Name) with or without last name prepositions. The default name format is the one suggested in the Biodiversity Information Standards (TDWG) format.

prepTDWG(
  x,
  sep = ", ",
  format = "last_init",
  pretty = TRUE,
  get.prep = FALSE,
  get.initials = TRUE,
  max.initials = 4
)

Arguments

x
the character string or vector containing the names.

sep
character. Input and output name separator. Default to ", ".

format
character. Output name format. The default is "last_init".

pretty
logical. Should the output name be returned in a pretty presentation (i.e. only the first letter of names capitalized, initials separated by points and no spaces, and family name prepositions in lower cases). Default to TRUE. If FALSE, names are returned in the same way as the input object x.

get.prep
logical. Should last name prepositions be included? Default to FALSE.

get.initials
logical. Should last name prepositions be included? Default to TRUE.

max.initials
numerical. Upper limit of number of letter for a single word to be considered as initials and not as a first name. Default to 4.

Value

The character string x in the standardized format.

Details

The default name format follows the one suggested by the TDWG, which is: Last name, followed by a comma and then the initials, separated by points (e.g. Hatschbach, G.G.).

The functions uses internally another plantR function: lastName(). So, it assumes that people last names are the ones provided at the end of the name string or preceding the name separator (i.e. comma), if present.

The function deals with simples last names, as well as with compound last names and last names with common name prefixes or prepositions (e.g. de, dos, van, ter, ...). By default, these prefixes and prepositions are removed, but they can be returned if the argument get.prep is set to TRUE.

The function assumes that all names containing separators (default to a comma) are in the format suggested by TDWG. But even for those cases, the function fixes simple problems (e.g. missing points between name initials).

If only one name is given, the function return the same name with the first letter capitalized.

The function output it is relatively stable regarding the input format, lower/uppercasing and spacing. But if the name provided has unusual formatting or if names for multiple people are provided within the same string, the function may not work properly. So, the output may depend on the input format and some level of double-checking may be necessary. See examples below.

References

Conn, Barry J. (ed.) (1996). HISPID 3 - Herbarium Information Standards and Protocols for Interchange of Data. Herbarium Information Systems Committee' (HISCOM). https://www.tdwg.org/standards/hispid3/

Willemse, L.P., van Welzen, P.C. & Mols, J.B. (2008). Standardisation in data-entry across databases: Avoiding Babylonian confusion. Taxon 57(2): 343-345.

See also

Author

Renato A. F. de Lima

Examples

  # Single names
  prepTDWG("gentry")

#> [1] "Gentry"

  prepTDWG("GENTRY")

#> [1] "Gentry"

  # Simple names
  prepTDWG("Alwyn Howard Gentry")

#> [1] "Gentry, A.H."

  prepTDWG("Alwyn H. Gentry")

#> [1] "Gentry, A.H."

  prepTDWG("A.H. Gentry")

#> [1] "Gentry, A.H."

  prepTDWG("A H Gentry")

#> [1] "Gentry, A.H."

  prepTDWG("Gentry, Alwyn Howard")

#> [1] "Gentry, A.H."

  prepTDWG("Gentry, AH")

#> [1] "Gentry, A.H."

  prepTDWG("Gentry AH")

#> [1] "Gentry, A.H."

  prepTDWG("GENTRY, A H")

#> [1] "Gentry, A.H."

  prepTDWG("gentry, alwyn howard")

#> [1] "Gentry, A.H."

  prepTDWG("gentry, a.h.")

#> [1] "Gentry, A.H."

  prepTDWG("gentry, a. h.")

#> [1] "Gentry, A.H."

  # Name with prepositions
  prepTDWG("Carl F. P. von Martius")

#> [1] "Martius, C.F.P."

  prepTDWG("Carl F. P. von Martius", get.prep = TRUE)

#> [1] "Martius, C.F.P. von"

  # Names with generational suffixes
  prepTDWG("Hermogenes de Freitas Leitao Filho")

#> [1] "Leitao Filho, H.F."

  prepTDWG("H.F. Leitao Filho")

#> [1] "Leitao Filho, H.F."

  prepTDWG("Leitao Filho, HF")

#> [1] "Leitao Filho, H.F."

  prepTDWG("Leitao filho, H. F.")

#> [1] "Leitao Filho, H.F."

  # Compound last name
  prepTDWG("Augustin Saint-Hilaire")

#> [1] "Saint-Hilaire, A."

  prepTDWG("A. Saint-Hilaire")

#> [1] "Saint-Hilaire, A."

  prepTDWG("Saint-Hilaire, Augustin")

#> [1] "Saint-Hilaire, A."

  # Other formats
  prepTDWG("John MacDonald")

#> [1] "MacDonald, J."

  prepTDWG("John McDonald")

#> [1] "McDonald, J."

  prepTDWG("John O'Brien")

#> [1] "O'Brien, J."

  # Multiple names, different settings
  names <- c("Gentry, AH", "Gentry A.H.",
  "Carl F. P. von Martius","Leitao filho, H. de F.",
  "Auguste de Saint-Hilaire", "John O'Reilly")
  prepTDWG(names)

#> [1] "Gentry, A.H."       "Gentry, A.H."       "Martius, C.F.P."   
#> [4] "Leitao Filho, H.F." "Saint-Hilaire, A."  "O'Reilly, J."      

  prepTDWG(names, format = "init_last")

#> [1] "A.H. Gentry"       "A.H. Gentry"       "C.F.P. Martius"   
#> [4] "H.F. Leitao Filho" "A. Saint-Hilaire"  "J. O'Reilly"      

  prepTDWG(names, format = "init_last", get.prep = TRUE)

#> [1] "A.H. Gentry"         "A.H. Gentry"         "C.F.P. von Martius" 
#> [4] "H.F. Leitao Filho"   "A. de Saint-Hilaire" "J. O'Reilly"        

  prepTDWG(names, get.prep = TRUE, format = "prep_last_init")

#> [1] "Gentry, A.H."         "Gentry, A.H."         "von Martius, C.F.P." 
#> [4] "Leitao Filho, H.F."   "de Saint-Hilaire, A." "O'Reilly, J."        

  prepTDWG(names, get.prep = TRUE, format = "prep_last_init",
           get.initials = FALSE)

#> [1] "Gentry, A.H."              "Gentry, A.H."             
#> [3] "von Martius, Carl F.P."    "Leitao Filho, H.F."       
#> [5] "de Saint-Hilaire, Auguste" "O'Reilly, John"           

  prepTDWG(names, get.prep = TRUE, pretty = FALSE,
           get.initials = FALSE)

#> [1] "Gentry, AH"                "Gentry, A. H."            
#> [3] "Martius, Carl F. P. von"   "Leitao filho, H. F."      
#> [5] "Saint-Hilaire, Auguste de" "O'Reilly, John"           

  ## Unusual formatting (function won't work always...)
  # two or more people names: output incorrect (combine names of authors)
  prepTDWG("C. Mendonca Filho; F. da Silva")

#> [1] "Silva, C.M.F."

  # two or more names, separated by comma: output incorrect (combine names of authors)
  prepTDWG("A. Alvarez, A. Zamora & V. Huaraca")

#> [1] "A. Alvarez, A.Z."

  # one name, two commas: fails to get all names
  prepTDWG("Cesar Sandro, Esteves, F")

#> [1] "Cesar Sandro, E.F."

  #' one name, abbreviations in the start and end: fails to get all names
  prepTDWG("C.S. Esteves F.")

#> [1] "C., S.E.F."