Validate Record Locality Information

Compares the resolution of the locality information provided in the original record with the resolution retrieved from the gazetteer.

validateLoc(x, res.orig = "resol.orig", res.gazet = "resolution.gazetteer")

Arguments

x
a data frame containing the resolution of the original and the gazetteer locality information.resolution from the gazetteer.

res.orig
character. The name of the column containing the resolution of the locality information from the original records. Default to 'resol.orig'.

res.gazet
character. The name of the column containing the resolution of the locality information from which the gazetteer coordinates were obtained. Default to 'resolution.gazetteer'.

Value

the data frame x with a new column 'loc.check' with the result of the locality validation.

Details

The function simply compares if the original information provided with each specimen is found in the locality gazetteer and at which resolution (i.e. country, state, municipality). This comparison is strongly dependent on the completeness of the gazetteer used. Please remind that the gazetteer provided with plantR is strongly biased towards Latin America, particularly Brazil.

The records whose locality was not found in the gazetteer at the same resolution are flagged with a "check_...". For these records, the locality resolution is downgraded until a locality is found in the gazetteer. If even the country name is not found, then the locality is flagged as "no_info". Specimens retaining their resolution are flagged with a "ok_same_resolution" and those found at a better resolution are flagged with an "ok_" (e.g. 'ok_state2municipality': a record which had its original resolution at the ' 'stateProvince' level is now at the 'municipality' level).

The records flagged with the class "check_local.2municip." should not be seen as problematic, since the plantR gazetteer contain a list of county names that is much more comprehensive than the list of locality names.

Author

Renato A. F. de Lima

Examples

(df <- data.frame(
resol.orig = c("locality", "municipality","stateProvince","country"),
resolution.gazetteer = c("county", "county","country", "county"),
stringsAsFactors = FALSE))

#>      resol.orig resolution.gazetteer
#> 1      locality               county
#> 2  municipality               county
#> 3 stateProvince              country
#> 4       country               county

res <- validateLoc(df)

#> [1] "Locality resolution in the original data vs. edited data:"
#>               original
#> edited         country locality municipality stateProvince
#>   country            0        0            0             1
#>   municipality       1        1            1             0

res

#>      resol.orig resolution.gazetteer             loc.check
#> 1      locality               county check_local.2municip.
#> 2  municipality               county    ok_same_resolution
#> 3 stateProvince              country   check_state2country
#> 4       country               county   ok_country2municip.